% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASURATDB.R
\docType{package}
\name{ASURATDB}
\alias{ASURATDB}
\title{ASURAT database builder}
\description{
ASURATDB is a software for collecting databases used by ASURAT.
   Using ASURATDB, one can import existing databases and create custom-built
   databases, which can be used by ASURAT. Examples include Disease Ontology,
   Cell Ontology, Gene Ontology, Kyoto Encyclopedia of Genes and Genomes
   (KEGG), Molecular Signatures Database (MSigDB).
}
