% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{collect_CO}
\alias{collect_CO}
\title{Collect a Cell Ontology database.}
\usage{
collect_CO(orgdb = NULL)
}
\arguments{
\item{orgdb}{A genome annotation package.}
}
\value{
Results from getCellOnto().
}
\description{
This function performs cellTypeToGenes() for collecting a Cell Ontology
  database.
}
\examples{
\dontrun{
# Below may be a time-consuming run.
dict_CO <- collect_CO(orgdb = org.Hs.eg.db::org.Hs.eg.db)
}

}
