% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{collect_GO}
\alias{collect_GO}
\title{Collect Gene Ontology database.}
\usage{
collect_GO(orgdb = NULL)
}
\arguments{
\item{orgdb}{A genome annotation package.}
}
\value{
Results from groupGO().
}
\description{
This function collects Gene Ontology database.
}
\examples{
\dontrun{
# Below may be a time-consuming run.
dict_GO <- collect_GO(orgdb = org.Hs.eg.db::org.Hs.eg.db)
}

}
