% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{compute_IC_CO}
\alias{compute_IC_CO}
\title{Compute information contents.}
\usage{
compute_IC_CO(dict = NULL, tidy = NULL, treeTable = NULL)
}
\arguments{
\item{dict}{A result of format_CO().}

\item{tidy}{A list of result of format_CO().}

\item{treeTable}{A result of make_treeTable_CO().}
}
\value{
A list of result of format_CO().
}
\description{
This function computes information contents for Cell Ontology terms.
}
\seealso{
Mistry and Pavlidis, BMC Bioinformatics, 2008.
}
