% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{do_cellTypeToGenes}
\alias{do_cellTypeToGenes}
\title{Do cellTypeToGenes().}
\usage{
do_cellTypeToGenes(data = NULL, orgdb = NULL)
}
\arguments{
\item{data}{Description.}

\item{orgdb}{A genome annotation package.}
}
\value{
Results of cellTypeToGenes().
}
\description{
This function performs cellTypeToGenes().
}
