% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{do_groupGO}
\alias{do_groupGO}
\title{Do groupGO().}
\usage{
do_groupGO(genes = NULL, orgdb = NULL, ont = NULL, level = NULL)
}
\arguments{
\item{genes}{A gene set.}

\item{orgdb}{A genome annotation package.}

\item{ont}{Category of Gene Ontology.}

\item{level}{Level of Gene Ontology.}
}
\value{
Results of groupGO().
}
\description{
This function performs groupGO().
}
