% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{do_keggGet}
\alias{do_keggGet}
\title{Do keggGet().}
\usage{
do_keggGet(data = NULL)
}
\arguments{
\item{data}{KEGG identifiers.}
}
\value{
Results of keggGet().
}
\description{
This function performs keggGet().
}
