% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{find_descendants}
\alias{find_descendants}
\title{Find all the descendants for parent terms.}
\usage{
find_descendants(id = NULL, co = NULL, map = NULL)
}
\arguments{
\item{id}{ID.}

\item{co}{A result of getCellOnto().}

\item{map}{A map data.}
}
\value{
A map data.
}
\description{
This function finds all the descendants for parent terms.
}
