% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_DO}
\alias{format_DO}
\title{Reformat a Disease Ontology database.}
\usage{
format_DO(dict = NULL, all_geneIDs = NULL, orgdb = NULL)
}
\arguments{
\item{dict}{A res slot of enrichDO().}

\item{all_geneIDs}{All genes in the form of ENTREZ ID.}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats a Disease Ontology database.
}
\examples{
\dontrun{
# Below may be time-consuming runs.
library(DOSE)
data(DO2EG)
dict_DO <- enrichDO(unlist(DO2EG), ont = "DO", pvalueCutoff = 1,
                    pAdjustMethod = "BH", minGSSize = 0, maxGSSize = 1e+10,
                    qvalueCutoff = 1, readable = FALSE)
human_DO <- format_DO(dict = dict_DO@result, all_geneIDs = dict_DO@gene,
                      orgdb = org.Hs.eg.db::org.Hs.eg.db)
}

}
