% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_GO}
\alias{format_GO}
\title{Reformat the result of collect_GO().}
\usage{
format_GO(dict = NULL, orgdb = NULL)
}
\arguments{
\item{dict}{A result of collect_GO().}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats the result of collect_GO().
}
\examples{
\dontrun{
# Below may be time-consuming runs.
dict_GO <- collect_GO(orgdb = org.Hs.eg.db::org.Hs.eg.db)
human_GO <- format_GO(dict = dict_GO, orgdb = org.Hs.eg.db::org.Hs.eg.db)
}

}
