% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{format_KEGG}
\alias{format_KEGG}
\title{Reformat collect_KEGG().}
\usage{
format_KEGG(dict = NULL, orgdb = NULL)
}
\arguments{
\item{dict}{A result of collect_KEGG().}

\item{orgdb}{A genome annotation package.}
}
\value{
A formatted database.
}
\description{
This function reformats collect_KEGG().
}
\examples{
\dontrun{
# Below may be time-consuming runs.
dict_KEGG <- collect_KEGG(organism = "hsa", categories = c("pathway"),
                          timelag = 0.1)
human_KEGG <- format_KEGG(
  dict = list(pathway = dict_KEGG[["pathway"]][["success"]]),
  orgdb = org.Hs.eg.db::org.Hs.eg.db)
}

}
