% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_databases.R
\name{make_treeTable_CO}
\alias{make_treeTable_CO}
\title{Make a parent-child relation table for CO terms.}
\usage{
make_treeTable_CO(tidy = NULL)
}
\arguments{
\item{tidy}{A list of result of format_CO().}
}
\value{
Parent-child relation table.
}
\description{
This function makes a parent-child relation table for CO terms.
}
